+++ NEWS.md
@@ -1,3 +1,11 @@
+## RAnEnExtra 0.2.7
+
+- Added `RAnEnExtra::queryObservationsTime` and `RAnEnExtra::queryObservationsSpace`
+
+## RAnEnExtra 0.2.6
+
+- Added `pre_sort` for `RAnEnExtra::verifyRankHist` to achieve better performance
+
 ## RAnEnExtra 0.2.5
 
 - Added function `parseProfiling`